// src/enums/massege.text.enum.ts
export enum MessageText {
  Start = "/start",
  Get = "🗳Ovoz berish",
  Back = "⬅️Orqaga",
  Balans = "💰Balans",
  Oplata = "💳 Yechib olish",
  Help = "💡Yordam",
  Manual = "📝Qo`llanma",
  To_Bot = "🤖Bot orqali",
  Tg_Users = "👤Telegram lichka orqali",
  Manual2 = "📝Batafsil ma`lumot"
}