// src/enums/texts.enum.ts
export enum Texts {
  StartText = "👇 Quyidagi bo'limlardan birini tanlang:\n\nImkon qadar lichkaga raqam yuboring!",
  NumberGet = "Shu ssilka orqali  ovoz bergandan song tasdiqlangan ovozni screenshot qilib botga tashlang shundan so`ng bir saot ichida hisobingizga pul tushadi\n\nO`QING❗️❗️❗️\nBot dan foydalanishdan oldin pastdagi 📝Qo`llanma tugmasi orqali o`tib foydalanish qo`llanmasini ko`rib chiqing❗️❗️❗️\n\n👇👇👇👇👇👇👇👇👇👇👇👇👇👇👇👇\nhttps://openbudget.uz/boards-list/2/9c061567-92f9-4181-a1c2-e582b4a95410",
  Oplata = "💳 Telefon yoki karta raqamingizni kiriting:\n\n⚠️ Bot Humans nomerga to'lov qilmaydi.\n\n📄 Na'muna:\n📱 Telefon raqam: +998912345678\n💳 Karta raqam: 8600111122223333'",
  Manual2 = " OpenBudjet'ning \"Mening yo'lim\" loyihasi doirasida qishlog'imiz ko'chalarining obod bo'lishida bizga ko'maklashing.\n\nHar bir ovoz uchun 3 000 so'm dan paynet.\n\nBu tashabbus 27-apreldan boshlandi, oldingi OpenBudjetga ovoz berganlar ham qayta ovoz berishlari mumkin.\n\nOvozlarni bot orqali yoki lichkaga yozish orqali berishingiz mumkin, to'lovlar o'sha zahoti amalga oshiriladi.\n\nBarchaga omad!",
  Tg_Users = "Quyidagi lichkaga shunchaki raqam yuboring, kodi kelsa kodni yuboring va paynetga ega bo'ling\n\nDiqqat raqam yuborgan vaqtingiz online holatda bo'ling, sms kodni vaqti 1 daqiqa, keyin yaroqsiz bo'lib qoladi.\n\n1-admin:\n\n2-admin:\n\n3-admin:\n\n👆 Yuqoridagi lichkalardan biriga raqam yuboring",
}

// OpenBudjet'ning "Mening yo'lim" loyihasi doirasida qishlog'imiz ko'chalarining obod bo'lishida bizga ko'maklashing.

// Har bir ovoz uchun 3 000 so'm dan paynet.

// Bu tashabbus 27-apreldan boshlandi, oldingi OpenBudjetga ovoz berganlar ham qayta ovoz berishlari mumkin.

// Ovozlarni bot orqali yoki lichkaga yozish orqali berishingiz mumkin, to'lovlar o'sha zahoti amalga oshiriladi.

// Barchaga omad!